Shuffle some code about
Also fix cyclop, funlen, gochecknoglobals, gochecknoinits, gocognit &
gocyclo lint warnings.
