+++ .golangci.yaml
@@ -2,13 +2,7 @@
 linters:
   enable-all: true
   disable:
-    - cyclop
     - exhaustivestruct
-    - funlen
-    - gochecknoglobals
-    - gochecknoinits
-    - gocognit
-    - gocyclo
     - goerr113
     - gomnd
     - godox
